backend/requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
dill
pycaret
catboost
xgboost
imbalanced-learn
django
djangorestframework
asgiref
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv

# -e .
